pyproject.toml
[project]
name = "prompt-sentinel"
version = "0.1.0"
description = "LLM Prompt Injection Detection Microservice - A defensive security tool for detecting and mitigating prompt injection attacks"
authors = [
    { name = "PromptSentinel Team", email = "admin@promptsentinel.ai" }
]
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Elastic-2.0" }
keywords = ["security", "llm", "prompt-injection", "ai-safety", "detection"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "License :: Other/Proprietary License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "websockets>=12.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "anthropic>=0.18.0",
    "openai>=1.12.0",
    "google-generativeai>=0.3.0",
    "redis>=5.0.0",
    "opentelemetry-api>=1.22.0",
    "opentelemetry-sdk>=1.22.0",
    "opentelemetry-instrumentation-fastapi>=0.43b0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
    "python-multipart>=0.0.6",
    "orjson>=3.9.0",
    "structlog>=24.1.0",
    "scikit-learn>=1.3.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "aiosqlite>=0.19.0",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
ml = [
    "hdbscan>=0.8.33",
    "sentence-transformers>=2.2.0",
    "umap-learn>=0.5.4",
]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "pytest-xdist>=3.5.0",
    "black>=23.12.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "types-pyyaml>=6.0.0",
    "types-redis>=4.6.0",
    "ipython>=8.0.0",
    "pre-commit>=3.6.0",
    "bandit[toml]>=1.7.0",
    "safety>=2.3.0",
]

test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "pytest-xdist>=3.5.0",
    "faker>=22.0.0",
    "hypothesis>=6.90.0",
    "aiohttp>=3.9.0",
    "asyncpg>=0.29.0",
    "motor>=3.3.0",
]

[project.urls]
"Homepage" = "https://github.com/promptsentinelai/prompt-sentinel"
"Bug Tracker" = "https://github.com/promptsentinelai/prompt-sentinel/issues"
"Documentation" = "https://github.com/promptsentinelai/prompt-sentinel/wiki"

[build-system]
requires = ["setuptools>=69.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "S",    # flake8-bandit (security)
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "S101",  # use of assert detected
    "S104",  # hardcoded bind to all interfaces (0.0.0.0) - acceptable for containers
    "S105",  # hardcoded password in string literal (false positive for enum values)
    "S108",  # hardcoded temp file paths in tests
    "S110",  # try-except-pass (acceptable for error recovery testing)
    "S311",  # random module for non-cryptographic use (tests and experiments)
    "S324",  # insecure hash function (MD5 for non-cryptographic hashing in experiments)
    "E402",  # module level import not at top of file (test files with skip markers)
    "S106",  # hardcoded password in function arguments (test fixtures)
    "S112",  # try-except-continue (acceptable for resilient testing)
]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = false
disallow_any_unimported = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = false
warn_no_return = true
check_untyped_defs = true
strict_equality = true
ignore_missing_imports = true
allow_untyped_calls = true
allow_untyped_decorators = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = [
    "-ra",
    "--strict-markers",
]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "e2e: marks tests as end-to-end tests",
    "performance: marks tests as performance tests",
    "docker: marks tests that require Docker (deselect with '-m \"not docker\"')",
]

[tool.coverage.run]
source = ["src/prompt_sentinel"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
